// src/api/generated/models/membershipControllerFindAllSortBy.ts
/**
 * Generated by orval v8.9.0 🍺
 * Do not edit manually.
 * Flowforge API
 * The Flowforge workflow orchestration engine API documentation
 * OpenAPI spec version: 1.0
 */

export type MembershipControllerFindAllSortBy = typeof MembershipControllerFindAllSortBy[keyof typeof MembershipControllerFindAllSortBy];


export const MembershipControllerFindAllSortBy = {
  name: 'name',
  createdAt: 'createdAt',
  role: 'role',
} as const;