// src/api/generated/models/membershipControllerFindAllSortOrder.ts
/**
 * Generated by orval v8.9.0 🍺
 * Do not edit manually.
 * Flowforge API
 * The Flowforge workflow orchestration engine API documentation
 * OpenAPI spec version: 1.0
 */

export type MembershipControllerFindAllSortOrder = typeof MembershipControllerFindAllSortOrder[keyof typeof MembershipControllerFindAllSortOrder];


export const MembershipControllerFindAllSortOrder = {
  asc: 'asc',
  desc: 'desc',
} as const;